Adding appsettings configuration, some code refactoring
