+++ app/backend/Extensions/KeyVaultConfigurationBuilderExtensions.cs
@@ -4,9 +4,10 @@ namespace MinimalApi.Extensions;
 
 internal static class KeyVaultConfigurationBuilderExtensions
 {
-    internal static IConfigurationBuilder ConfigureAzureKeyVault(this IConfigurationBuilder builder)
+    internal static ConfigurationManager ConfigureAzureKeyVault(this ConfigurationManager builder)
     {
-        var azureKeyVaultEndpoint = Environment.GetEnvironmentVariable("AZURE_KEY_VAULT_ENDPOINT");
+        var azureKeyVaultEndpoint  = builder["AZURE_KEY_VAULT_ENDPOINT"];
+        //var azureKeyVaultEndpoint = Environment.GetEnvironmentVariable("AZURE_KEY_VAULT_ENDPOINT");
         ArgumentNullException.ThrowIfNullOrEmpty(azureKeyVaultEndpoint);
 
         builder.AddAzureKeyVault(
