+++ app/backend/appsettings.json
@@ -5,5 +5,11 @@
       "Microsoft.AspNetCore": "Warning"
     }
   },
-  "AllowedHosts": "*"
+  "AllowedHosts": "*",
+  "AZURE_KEY_VAULT_ENDPOINT": "https://az900-keyvault-test.vault.azure.net/",
+  "AzureOpenAiServiceEndpoint": "https://pztestopenainew.openai.azure.com/",
+  "AzureOpenAiApiKey": "846ce93409d0430abaf5448f2913566f",
+  "AzureOpenAiChatGptDeployment": "gpt-35-turbo",
+  "AzureStorageAccountEndpoint": "https://stpztestaire321629269342.blob.core.windows.net",
+  "AzureStorageContainer": "az-ai-search"
 }
